imagedatabase as own thread
